added molset argument, keep job logic, removed old code

+++ package/chemicalchecker/util/pipeline/tasks/task_cc_fit_predict.py
@@ -1,11 +1,11 @@
 """CCFit task.
 
-This task allow submitting HPC jobs which will 'fit' for a specific signature
-type (e.g. 'sign1').
+This task allow submitting HPC jobs which will call the `fit` mwthod
+for a specific signature type (e.g. 'sign1').
 It allows submitting jobs for all spaces of the CC at once and passing
 parameters specific for each of them.
-We should avoid adding too much logic in here, and simply pass args and
-kwargs to signatures. The specific signature classes should check that
+We should avoid adding too much logic in here, and simply pass `args` and
+`kwargs` to signatures. The specific signature classes should check that
 the parameters are OK.
 """
 import os
@@ -20,70 +20,7 @@ from chemicalchecker.core import ChemicalChecker
 from chemicalchecker.util.pipeline import BaseTask
 from chemicalchecker.util import logged, Config, HPC
 
-
-VALID_TYPES = ['sign', 'neig', 'clus', 'proj', 'diag']
-
-DEPENDENCIES = {
-    'sign0': ['sign0'],
-    'sign1': ['sign0'],
-    'sign2': ['sign1', 'neig1'],
-    'sign3': ['sign2'],
-    'neig1': ['sign1'],
-    'neig2': ['sign2'],
-    'neig3': ['sign3'],
-    'clus1': ['sign1'],
-    'clus2': ['sign2'],
-    'clus3': ['sign3'],
-    'proj1': ['sign1'],
-    'proj2': ['sign2'],
-    'proj3': ['sign3']
-}
-
-HPC_PARAMS = {
-    'sign0': {'memory': 44, 'cpu': 22},
-    'sign1': {'memory': 40, 'cpu': 10},
-    'sign2': {'memory': 20, 'cpu': 16},
-    'sign3': {'memory': 2,  'cpu': 32},
-    'neig1': {'memory': 30, 'cpu': 15},
-    'neig2': {'memory': 30, 'cpu': 15},
-    'neig3': {'memory': 30, 'cpu': 15},
-    'clus1': {'memory': 20, 'cpu': 10},
-    'clus2': {'memory': 20, 'cpu': 10},
-    'clus3': {'memory': 20, 'cpu': 10},
-    'proj1': {'memory': 20, 'cpu': 10},
-    'proj2': {'memory': 20, 'cpu': 10},
-    'proj3': {'memory': 20, 'cpu': 10}
-}
-
-SIGN_PARAMS = {
-    'sign2': {'adanet': {'cpu': 16}, 'node2vec': {'cpu': 4}},
-    'sign3': {'cpu': 32},
-    'neig1': {'cpu': 15},
-    'neig2': {'cpu': 15},
-    'neig3': {'cpu': 15},
-    'clus1': {'cpu': 10},
-    'clus2': {'cpu': 10},
-    'clus3': {'cpu': 10},
-    'proj1': {'cpu': 10},
-    'proj2': {'cpu': 10},
-    'proj3': {'cpu': 10}
-}
-
-MOLSET_FIT = {
-    'sign0': 'full',
-    'sign1': 'full',
-    'sign2': 'reference',
-    'sign3': 'full',
-    'neig1': 'reference',
-    'neig2': 'reference',
-    'neig3': 'reference',
-    'clus1': 'reference',
-    'clus2': 'reference',
-    'clus3': 'reference',
-    'proj1': 'reference',
-    'proj2': 'reference',
-    'proj3': 'reference'
-}
+VALID_TYPES = ['sign', 'neig', 'clus', 'proj']
 
 FIT_SCRIPT = """
 import sys
@@ -109,98 +46,18 @@ sign.fit(*fit_args, **fit_kwargs)
 print('JOB DONE')
 """
 
-"""
-CC_SCRIPT_FR = [
-    'sign_new_ref = cc.get_signature("<CCTYPE>", "reference", ds, **sign_pars)',
-    "sign_new_ref.fit(sign_ref, **fit_pars)",
-    "sign_new_full = cc.get_signature('<CCTYPE>', 'full', ds, **sign_pars)",
-    "sign_new_ref.predict(sign_full, destination=sign_new_full.data_path)",
-    "sign_new_full.mark_ready()"
-]
-
-CC_SCRIPT_F = [
-    'sign_new_full = cc.get_signature("<CCTYPE>", "full", ds, **sign_pars)',
-    "sign_new_full.fit(sign_full)"
-    "if sign_new_full.cctype.startswith('sign'):",
-    '    diag=cc.diagnosis(sign_new_full)',
-    '    diag.canvas()'
-]
-
-SIGN2_SCRIPT_FR = [
-    'neig1_ref = cc.get_signature("neig1", "reference", ds)',
-    'sign2_ref = cc.get_signature("sign2", "reference", ds, **sign_pars)',
-    "sign2_ref.fit(sign_ref, neig1_ref, reuse=False)",
-    "sign2_full = cc.get_signature('sign2', 'full', ds, **sign_pars)",
-    "sign2_ref.predict(sign_full, destination=sign2_full.data_path)",
-    "sign2_full.validate()",
-    "sign2_full.mark_ready()"
-]
-
-SIGN2_SCRIPT_F = [
-    'neig1_full = cc.get_signature("neig1", "full", ds)',
-    'sign2_full = cc.get_signature("sign2", "full", ds, **sign_pars)',
-    "sign2_full.fit(sign_full, neig1_full, reuse=False)"
-]
-
-SIGN3_SCRIPT_F = [
-    "sign1_full = cc.get_signature('sign1', 'full', ds, **sign_pars)",
-    "sign3_full = cc.get_signature('sign3', 'full', ds, **sign_pars)",
-    "sign2_list = [cc.get_signature('sign2', 'full', ds) for ds in sign2_src_list]",
-    "sign2_list.append(cc.get_signature('sign2', 'full', ds))",
-    "sign3_full.fit(sign2_list, sign_full, sign1_full)",
-    'diag=cc.diagnosis(sign3_full)',
-    'diag.canvas()'
-]
-
-SIGN0_SCRIPT_FR = [
-    "if len(fit_pars) == 0:",
-    "    prepro_file = cc.preprocess(sign_full)",
-    "    cc_old = ChemicalChecker(CC_OLD_PATH)",
-    "    pars['data_file'] = prepro_file",
-    "    pars['cc'] = cc_old",
-    "sign_full.fit(**fit_pars)",
-    'diag=cc.diagnosis(sign_full)',
-    'diag.canvas()'
-]
-
-SIGN1_SCRIPT_FR = [
-    'sign_new_full = cc.get_signature("sign1", "full", ds)',
-    'sign_new_full.fit(sign_full, **fit_pars)'
-    'diag=cc.diagnosis(sign_new_full)',
-    'diag.canvas()'
-]
-
-
-SPECIFIC_SCRIPTS = {
-    'sign1': (SIGN1_SCRIPT_FR, SIGN1_SCRIPT_FR),
-    'sign2': (SIGN2_SCRIPT_FR, SIGN2_SCRIPT_F),
-}
-
-
-SIGN3_SCRIPT_FR = [
-    "sign1_full = cc.get_signature('sign1', 'full', data, **sign_pars)",
-    "sign3_full = cc.get_signature('sign3', 'full', data, **sign_pars)",
-    "sign2_src_list = [%s]" % (str(self.ref_datasets)[1:-1]),
-    "sign2_list = [cc.get_signature('sign2', 'full', ds) for ds in sign2_src_list]",
-    "sign3_full.fit(sign2_list,sign_full,sign1_full,sign2_universe='%s', sign2_coverage='%s')" % (
-        full_universe, full_coverage)
-]
-SPECIFIC_SCRIPTS['sign3'] = (
-    SIGN3_SCRIPT_FR, ["sign2_src_list = [%s]" %
-                      (str(self.ref_datasets)[1:-1])] + SIGN3_SCRIPT_F)
-
-"""
-
 
 @logged
 class CCFit(BaseTask):
 
-    def __init__(self, cc_root, cctype, **params):
+    def __init__(self, cc_root, cctype, molset, **params):
         """Initialize CC fit task.
 
         Args:
             cc_root (str): The CC root path (Required)
             cctype (str): The CC type where the fit is applied (Required)
+            molset (str): The signature molset (e.g. `full` or `reference`)
+                on which the `fit` method will be called.
             name (str): The name of the task (default:cctype)
             datasets (list): The list of dataset codes to apply the fit
                 (Optional, by default 'essential' which includes all essential
@@ -232,6 +89,7 @@ class CCFit(BaseTask):
         BaseTask.__init__(self, self.name)
         self.cctype = cctype
         self.cc_root = cc_root
+        self.molset = molset
         self.datasets = params.get('datasets', 'essential')
         if self.datasets == 'essential':
             self.datasets = [ds.code for ds in Dataset.get(essential=True)]
@@ -250,17 +108,13 @@ class CCFit(BaseTask):
         cc = ChemicalChecker(self.cc_root)
         dataset_codes = list()
         for ds in self.datasets:
-            sign = cc.get_signature(self.cctype, MOLSET_FIT[self.cctype], ds)
+            sign = cc.get_signature(self.cctype, self.molset, ds)
             if sign.is_fit():
                 continue
-            # special exception to avoid recomputing from the beginning
-            # sign0 of D1.001, even if not completed
-            if os.path.exists(sign.signature_path):
-                if ds == 'D1.001' and self.cctype == 'sign0':
-                    continue
             dataset_codes.append(ds)
         if len(dataset_codes) == 0:
             self.__log.warning('All dataset are already fitted.')
+            self.mark_ready()
             return
 
         # sign3 specific pre-calculations
@@ -302,7 +156,7 @@ class CCFit(BaseTask):
             # we add arguments which are used by CCFit but are also needed
             # by a signature fit
             sign_args.insert(0, self.cctype)
-            sign_args.insert(1, MOLSET_FIT[self.cctype])
+            sign_args.insert(1, self.molset)
             sign_args.insert(2, ds_code)
             # prepare it as tuple that will be serialized
             dataset_params.append(
@@ -346,7 +200,7 @@ class CCFit(BaseTask):
         dataset_not_done = []
         for ds_code in dataset_codes:
             sign = cc.get_signature(
-                self.cctype, MOLSET_FIT[self.cctype], ds_code)
+                self.cctype, self.molset, ds_code)
             if sign.is_fit():
                 continue
             dataset_not_done.append(ds_code)
@@ -355,8 +209,9 @@ class CCFit(BaseTask):
 
         if len(dataset_not_done) == 0:
             self.mark_ready()
-            self.__log.info("Deleting job path: %s", job_path)
-            shutil.rmtree(job_path, ignore_errors=True)
+            if not self.keep_jobs:
+                self.__log.info("Deleting job path: %s", job_path)
+                shutil.rmtree(job_path, ignore_errors=True)
         else:
             if not self.custom_ready():
                 raise Exception("Not all dataset fits are done")
